requirements.txt
requests>=2.32.4
python-dotenv>=1.1.1
tqdm>=4.67.1
backoff>=2.2.1 